requirements.txt
absl-py==2.3.0
addict==2.4.0
basicsr==1.4.2
certifi==2025.7.9
charset-normalizer==3.4.2
conda-pack @ file:///home/conda/feedstock_root/build_artifacts/conda-pack_1734292804024/work
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.13
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.58.5
fsspec==2024.6.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.73.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
Jinja2==3.1.4
joblib==1.5.1
jupyter_client==8.6.3
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.7.2
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.1
realesrgan==0.3.0
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sympy==1.13.3
tb-nightly==2.20.0a20250711
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
torch==2.7.1+cu126
torchaudio==2.7.1+cu126
torchvision==0.22.1+cu126
tornado==6.4.2
tqdm==4.32.2
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
yapf==0.43.0
